requirements.txt
requests
python-dotenv
spotipy
PyGithub